moved NN and Canvas logic to separate component add nav/menu
