+++ src/components/App/App.module.css
@@ -1,50 +1,50 @@
-@keyframes rotate {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
+.root {
+  width: 100vw;
+  height: 100vh;
+  overflow: hidden;
+  position: relative;
+  font-family: 'Roboto Mono', sans-serif;
 }
 
-.canvasContainer {
+.app {
   position: absolute;
-  top: 30px;
-  left: 30px;
-  bottom: 30px;
-  right: 30px;
+  padding: 0 60px;
+  box-sizing: border-box;
 }
 
-.canvasScaleContainer {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  width: 100%;
-  height: 100%;
-  box-shadow: 1px 3px 6px rgba(0, 0, 0, 0.33);
-  touch-action: none;
+.header {
+  height: 60px;
+  display: flex;
+  align-items: center;
+  flex-wrap: nowrap;
 }
 
-.canvasContainer canvas {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  top: 0;
-  left: 0;
+.headerMenu {
+  flex: 1 1 auto;
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
 }
 
-.spinner {
-  position: absolute;
-  bottom: 40px;
-  right: 40px;
-  z-index: 2;
-  animation: rotate 1s linear infinite;
-  width: 30px;
+.headerButton {
+  font-family: 'Roboto Mono', sans-serif;
+  font-size: 15px;
+  cursor: pointer;
+  border: 0;
+  background: transparent;
+  padding: 0;
+  color: #666;
+  text-transform: uppercase;
+  margin: 0 15px;
   height: 30px;
-  opacity: 0;
-  transition: opacity 0.3s linear;
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
 }
 
-.spinnerVisible {
-  opacity: 1;
+.headerButton:first-child {
+  margin-left: 0;
+}
+.headerButton:last-child {
+  margin-right: 0;
 }
