+++ src/settings.ts
@@ -1,0 +1,3 @@
+export const CANVAS_WIDTH = 512;
+export const CANVAS_HEIGHT = 256;
+export const CANVAS_GUTTER = 60;
